requirements.txt
asgiref==3.8.1
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.42.1
cryptography==44.0.0
Django==5.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-webpack-loader==3.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
firebase-admin==6.6.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
gunicorn==23.0.0
httplib2==0.22.0
idna==3.10
msgpack==1.1.0
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
